main/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baofeng-logo-flasher"
version = "0.1.0"
description = "Standalone logo flasher for Baofeng UV-5RM radios with safety verification"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Codex", email = "codex@localhost.local"}]
license = {text = "MIT"}

dependencies = [
    "pyserial>=3.5",
    "pillow>=10.0.0",
    "rich>=13.0.0",
    "typer[all]>=0.9.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.28.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
baofeng-logo-flasher = "baofeng_logo_flasher.cli:main"
baofeng-logo-flasher-ui = "baofeng_logo_flasher.streamlit_ui:launch"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["src"]
